+++ dashboard/src/Router.tsx
@@ -43,28 +43,37 @@ import LibraryBlogCourse from "./pages/library/blog/course";
 import LibraryBlogAnthology from "./pages/library/blog/anthology";
 import LibraryBlogTerm from "./pages/library/blog/term";
 
-import StudioHome from "./pages/studio";
+import Studio from "./pages/studio";
+import StudioHome from "./pages/studio/home";
 
 import StudioPalicanon from "./pages/studio/palicanon";
 import StudioRecent from "./pages/studio/recent";
 
 import StudioChannel from "./pages/studio/channel";
+import StudioChannelList from "./pages/studio/channel/list";
 import StudioChannelEdit from "./pages/studio/channel/edit";
 
 import StudioGroup from "./pages/studio/group";
+import StudioGroupList from "./pages/studio/group/list";
 import StudioGroupEdit from "./pages/studio/group/edit";
 import StudioGroupShow from "./pages/studio/group/show";
 
 import StudioDict from "./pages/studio/dict";
+import StudioDictList from "./pages/studio/dict/list";
+
 import StudioTerm from "./pages/studio/term";
+import StudioTermList from "./pages/studio/term/list";
 
 import StudioArticle from "./pages/studio/article";
+import StudioArticleList from "./pages/studio/article/list";
 import StudioArticleEdit from "./pages/studio/article/edit";
 
 import StudioAnthology from "./pages/studio/anthology";
+import StudioAnthologyList from "./pages/studio/anthology/list";
 import StudioAnthologyEdit from "./pages/studio/anthology/edit";
 
 import StudioAnalysis from "./pages/studio/analysis";
+import StudioAnalysisList from "./pages/studio/analysis/list";
 
 const Widget = () => {
 	return (
@@ -99,7 +108,6 @@ const Widget = () => {
 				<Route path="list" element={<LibraryCommunityList />} />
 				<Route path="recent" element={<LibraryCommunityRecent />} />
 			</Route>
-
 			<Route path="palicanon" element={<LibraryPalicanon />}>
 				<Route path="list" element={<LibraryPalicanonByPath />} />
 				<Route path="list/:root" element={<LibraryPalicanonByPath />} />
@@ -107,12 +115,10 @@ const Widget = () => {
 				<Route path="list/:root/:path/:tag" element={<LibraryPalicanonByPath />} />
 				<Route path="chapter/:id" element={<LibraryPalicanonChapter />} />
 			</Route>
-
 			<Route path="course" element={<LibraryCourse />}>
+				<Route path="list" element={<LibraryCourseList />}></Route>
 				<Route path="show/:id" element={<LibraryCourseShow />}></Route>
 				<Route path="lesson/:id" element={<LibraryLessonShow />}></Route>
-				<Route path="course/:id" element={<LibraryLessonShow />}></Route>
-				<Route path="list" element={<LibraryCourseList />}></Route>
 			</Route>
 
 			<Route path="term/:word" element={<LibraryTerm />} />
@@ -127,7 +133,13 @@ const Widget = () => {
 				<Route path=":id" element={<LibraryAnthologyShow />} />
 				<Route path=":id/by_channel/:tags" element={<LibraryAnthologyShow />} />
 			</Route>
-			<Route path="article/show/:id" element={<LibraryArticle />} />
+
+			<Route path="article" element={<LibraryArticle />}>
+				<Route path=":type/:id" element={<LibraryArticle />} />
+				<Route path=":type/:id/param/:param" element={<LibraryArticle />} />
+				<Route path=":type/:id/tran" element={<LibraryArticle />} />
+				<Route path=":type/:id/tran/param/:param" element={<LibraryArticle />} />
+			</Route>
 
 			<Route path="blog/:studio" element={<LibraryBlog />}>
 				<Route path="overview" element={<LibraryBlogOverview />} />
@@ -137,27 +149,37 @@ const Widget = () => {
 				<Route path="term" element={<LibraryBlogTerm />} />
 			</Route>
 
-			<Route path="studio/:studioname" element={<StudioHome />}></Route>
-			<Route path="studio/:studioname/palicanon" element={<StudioPalicanon />}></Route>
-			<Route path="studio/:studioname/recent" element={<StudioRecent />}></Route>
-
-			<Route path="studio/:studioname/channel" element={<StudioChannel />}></Route>
-			<Route path="studio/:studioname/channel/edit/:channelid" element={<StudioChannelEdit />} />
-
-			<Route path="studio/:studioname/group" element={<StudioGroup />}></Route>
-			<Route path="studio/:studioname/group/:groupid" element={<StudioGroupShow />} />
-			<Route path="studio/:studioname/group/edit/:groupid" element={<StudioGroupEdit />} />
-
-			<Route path="studio/:studioname/dict" element={<StudioDict />} />
-			<Route path="studio/:studioname/term" element={<StudioTerm />} />
-
-			<Route path="studio/:studioname/article" element={<StudioArticle />}></Route>
-			<Route path="studio/:studioname/article/edit/:articleid" element={<StudioArticleEdit />} />
-
-			<Route path="studio/:studioname/anthology" element={<StudioAnthology />}></Route>
-			<Route path="studio/:studioname/anthology/edit/:anthology_id" element={<StudioAnthologyEdit />} />
-
-			<Route path="studio/:studioname/analysis" element={<StudioAnalysis />}></Route>
+			<Route path="studio/:studioname" element={<Studio />}>
+				<Route path="home" element={<StudioHome />} />
+				<Route path="palicanon" element={<StudioPalicanon />}></Route>
+				<Route path="recent" element={<StudioRecent />}></Route>
+				<Route path="channel" element={<StudioChannel />}>
+					<Route path="list" element={<StudioChannelList />} />
+					<Route path=":channelid/edit" element={<StudioChannelEdit />} />
+				</Route>
+				<Route path="group" element={<StudioGroup />}>
+					<Route path="list" element={<StudioGroupList />} />
+					<Route path=":groupid" element={<StudioGroupShow />} />
+					<Route path=":groupid/edit" element={<StudioGroupEdit />} />
+				</Route>
+				<Route path="dict" element={<StudioDict />}>
+					<Route path="list" element={<StudioDictList />} />
+				</Route>
+				<Route path="term" element={<StudioTerm />}>
+					<Route path="list" element={<StudioTermList />} />
+				</Route>
+				<Route path="article" element={<StudioArticle />}>
+					<Route path="list" element={<StudioArticleList />} />
+					<Route path=":articleid/edit" element={<StudioArticleEdit />} />
+				</Route>
+				<Route path="anthology" element={<StudioAnthology />}>
+					<Route path="list" element={<StudioAnthologyList />}></Route>
+					<Route path=":anthology_id/edit" element={<StudioAnthologyEdit />} />
+				</Route>
+				<Route path="analysis" element={<StudioAnalysis />}>
+					<Route path="list" element={<StudioAnalysisList />} />
+				</Route>
+			</Route>
 		</Routes>
 	);
 };
