+++ dashboard/src/pages/studio/palicanon/index.tsx
@@ -1,16 +1,14 @@
 import { useParams, Link } from "react-router-dom";
 import { useIntl } from "react-intl";
 import { Space } from "antd";
-import HeadBar from "../../../components/studio/HeadBar";
+
 import LeftSider from "../../../components/studio/LeftSider";
-import Footer from "../../../components/studio/Footer";
 
 const Widget = () => {
 	const intl = useIntl(); //i18n
 	const { studioname } = useParams(); //url 参数
 	return (
 		<div>
-			<HeadBar />
 			<LeftSider />
 			<h2>
 				studio/{studioname}/{intl.formatMessage({ id: "columns.studio.palicanon.title" })}
@@ -20,7 +18,6 @@ const Widget = () => {
 					<Link to=""> </Link>
 				</Space>
 			</div>
-			<Footer />
 		</div>
 	);
 };
